chore: refactor a few utils
